Setup Poetry, Tox, Travis, and add PB800 as an example (#1)
This provides the main setup for this new repository, as follows:

## Poetry

Currently, we use Flit for Pex and Pipenv for `rbe-token-server`. However, we decided to switch over Pex and `rbe-token-server` to use Poetry. Why? Flit can't be used for `rbe-token-server`, which is an app, and we want to be able to use the same tool for all of our repos.

Poetry also has substantial community momentum currently.

## Tox

We use Tox to simplify testing every supported Python version and to align with Pex.

## Travis

We stick with Travis because all other projects use Travis and we're familiar with it.

## Flake8 plugin setup

This follows the architecture of [`flake8-2020`](https://github.com/asottile/flake8-2020/blob/master/flake8_2020.py) in its definition of a `Visitor(ast.NodeVisitor)` and `Plugin` class. This plugin is developed by one of the maintainers of Flake8 and is only 4 months old, so is a good example of how to write Flake8 plugins.  

After reviewing several plugins, including `flake8-2020`, [`flake8-bugbears`](https://github.com/PyCQA/flake8-bugbear), [`flake8-eradicate`](https://github.com/sobolevn/flake8-eradicate), [`flake8-print`](https://github.com/JBKahn/flake8-print), and [`flake8-comprehensions`](https://github.com/adamchainz/flake8-comprehensions), it appears Flake8 plugins conventionally define all their logic in one module and one `ast.NodeVisitor` class, in contrast to `pantsbuild.contrib...checker` having a distinct file for each check. If this ends up not scaling well, we'll revisit alternative architectures, but for now, we follow the convention of other projects.

### Testing

We use `pytest-flake8dir`, which is a utility for testing Flake8 plugins. This makes it easy for us to write mock files to a tempdir and then run Flake8 against them and to capture the output.