Add constant logic checks (PB804 and PB805) (#3)
These check for likely issues in boolean expressions. Using constants in a boolean expression means that the conditional will short-circuit, which is likely a bug.

MyPy does have some checking for this already, but MyPy does not complain about using truthy strings and numbers nor do we want to require users to start using MyPy to get the benefits of this lint.