Disable Python 2->3 lints by default (#11)
Now that Python 2 is EOL, Pants assumes that Python codebases are using Python 3. They can still use Python 2, but the user must opt-in to Python 2 support.

The author of Flake8 suggested a very nice design pattern in https://gitlab.com/pycqa/flake8/-/issues/482 to have an optional plugin as a part of the core distribution. All that users need to do is set `--enable-extensions PB6` to turn on this entire set of lints.

Once we add the formatting lints, like checking for 2-space indentation, we will be able to use this same pattern to have a FormattingPlugin, which is disabled by default.