Fix error messages getting printed 4 times (#18)
Turns out, subclassing `Plugin` 3 different times means that Flake8 calls `Plugin.run()` 4 times and generates an error 4 different times. We didn't detect this in tests because we use `sets` due to a Python 2 bug with our testing library.

The solution is to stop subclassing `Plugin` and treat `IndentationPlugin`, `TrailingSlashesPlugin`, and `SixPlugin` as distinct, first-class plugins.

This also results in cleaner code.